requirements/dev.txt
-r base.txt
django-debug-toolbar
ipdb
ipython
fabric3